// frontend/src/types/DoctorAppointment.ts
export interface DoctorAppointment {
    id: number
    patientNo: number
    patientName: string
    departmentName: string
    appointmentDate: any
    appointmentTime: any
    status: string
    notes: string
    patientEmail: string
    gender: string
    age: string
    bloodGroup: string
    phoneNumber: string
} 